add:Light, and some move ref

+++ 3DViewer/graph/LearnMesh.h
@@ -85,7 +85,7 @@ namespace rtx
 
             // draw mesh
             glBindVertexArray(VAO);
-            glDrawElements(GL_TRIANGLES, indices.size(), GL_UNSIGNED_INT, 0);
+            glDrawElements(GL_TRIANGLES, static_cast<int>(indices.size()), GL_UNSIGNED_INT, 0);
             glBindVertexArray(0);
 
             // always good practice to set everything back to defaults once configured.
